분리된 위젯들 파일 분리
